split out audio for organization
